refactor: Cameras folder
