+++ src/cameras/camera.cpp
