+++ src/cameras/camera.h
@@ -1,8 +1,8 @@
 #ifndef CAMERA_H
 #define CAMERA_H 1
 
-#include "rt3.h"
-#include "ray.h"
+#include "../core/rt3.h"
+#include "../core/ray.h"
 #include "film.h"
 
 namespace rt3{
