+++ src/cameras/film.h
@@ -1,9 +1,9 @@
 #ifndef FILM_H
 #define FILM_H
 
-#include "error.h"
-#include "paramset.h"
-#include "rt3.h"
+#include "../core/error.h"
+#include "../core/paramset.h"
+#include "../core/rt3.h"
 
 namespace rt3 {
